consolidating service/repository logic, removing unneeded data endpoint
